refactor: HttpRequest refactoring
